+++ CHANGELOG.md
@@ -8,6 +8,7 @@ for Rust libraries in [RFC #1105](https://github.com/rust-lang/rfcs/blob/master/
 
  * Updated `onig` to `6.3`.
  * Switched rust edition to 2021.
+ * `master` branch is now called `main`.
 
 ## [1.2.0] - 2021-03-21